# contracts/openapi/platform-api.v1.yaml
openapi: 3.1.0
info:
  title: VPP Platform API
  version: 1.0.0
  description: Control-plane API contract for the VPP platform.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /portfolios:
    get:
      operationId: listPortfolios
      tags: [Resources]
      summary: List portfolios in the authenticated tenant
      description: Returns only portfolios belonging to the authenticated tenant.
      responses:
        "200":
          description: Authorized portfolios
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Portfolio" } }
    post:
      operationId: createPortfolio
      tags: [Resources]
      summary: Create a tenant-scoped resource portfolio
      description: Requires TENANT_ADMIN and records an audit event in the same transaction.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePortfolioRequest" }
      responses:
        "201":
          description: Portfolio created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Portfolio" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /sites:
    post:
      operationId: createSite
      tags: [Resources]
      summary: Create a site inside an authorized portfolio
      description: The portfolio lookup is tenant scoped and an invalid time zone is rejected.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSiteRequest" }
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /device-models:
    post:
      operationId: createDeviceModel
      tags: [Resources]
      summary: Create a versioned device model and telemetry point schema
      description: Model names are unique by tenant and schema version.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeviceModelRequest" }
      responses:
        "201":
          description: Device model created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceModel" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /devices:
    get:
      operationId: listDevices
      tags: [Resources]
      summary: List devices in the authenticated tenant only
      description: Never returns devices belonging to another tenant.
      responses:
        "200":
          description: Tenant-scoped device list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Device" }
    post:
      operationId: createDevice
      tags: [Resources]
      summary: Register a device with capabilities
      description: A new device remains PENDING until an active credential is provisioned.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeviceRequest" }
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /devices/{deviceId}:
    get:
      operationId: getDevice
      tags: [Resources]
      summary: Get one tenant-scoped device
      description: Returns not found for both missing and cross-tenant device identifiers.
      parameters: [{ $ref: "#/components/parameters/DeviceId" }]
      responses:
        "200":
          description: Device found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "404": { $ref: "#/components/responses/NotFound" }
  /devices/{deviceId}/status:
    put:
      operationId: updateDeviceStatus
      tags: [Resources]
      summary: Change device eligibility status and increment configuration version
      description: ACTIVE requires a current credential; other states remove scheduling eligibility.
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDeviceStatusRequest" }
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /devices/{deviceId}/credentials:
    post:
      operationId: rotateDeviceCredential
      tags: [Resources]
      summary: Replace a device credential without returning or storing its plaintext
      description: Stores a keyed verifier and revokes the previous reference atomically.
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RotateCredentialRequest" }
      responses:
        "201":
          description: Credential verifier stored; plaintext is never returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialReference" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /tariff-plans:
    get:
      operationId: listTariffPlans
      tags: [Tariffs]
      summary: List versioned tariff plans in the authenticated tenant
      description: Returns tariff metadata for schedule input selection; interval prices remain frozen by the referenced plan version.
      responses:
        "200":
          description: Tariff plans ordered by creation time, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TariffPlan" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createTariffPlan
      tags: [Tariffs]
      summary: Create a versioned tariff whose UTC intervals cannot overlap
      description: Application validation and a PostgreSQL exclusion constraint both reject overlap.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTariffPlanRequest" }
      responses:
        "201":
          description: Tariff created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TariffPlan" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [Schedules]
      summary: List tenant-scoped schedule outcomes
      description: Returns validated candidates and explicit infeasible attempts, optionally filtered by portfolio.
      parameters:
        - { name: portfolioId, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Schedule outcomes
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Schedule" } } } }
    post:
      operationId: createSchedule
      tags: [Schedules]
      summary: Generate and hard-constraint-validate an immutable candidate schedule
      description: Returns a terminal VALIDATED candidate or an explicit INFEASIBLE result without a version.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "202":
          description: Schedule generation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleDetail"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /schedules/{scheduleId}:
    get:
      operationId: getSchedule
      tags: [Schedules]
      summary: Get one terminal schedule outcome and immutable version
      description: Returns the frozen input evidence, result summary, interval series and device targets when feasible.
      parameters:
        - $ref: "#/components/parameters/ScheduleId"
      responses:
        "200":
          description: Schedule found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleDetail"
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ScheduleId:
      name: scheduleId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key reused for retries of the same semantic write.
      schema:
        type: string
        minLength: 8
        maxLength: 200
  responses:
...
    Forbidden:
      description: Authenticated actor lacks the required role
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Object does not exist or is outside the authorized tenant scope
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: State or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: Request or domain validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    CreatePortfolioRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    Portfolio:
      type: object
      additionalProperties: false
      required: [id, name, status, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        created_at: { type: string, format: date-time }
    CreateSiteRequest:
      type: object
      additionalProperties: false
      required: [portfolio_id, name, timezone, grid_connection_limit_kw]
      properties:
        portfolio_id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        timezone: { type: string, minLength: 1, maxLength: 80 }
        grid_connection_limit_kw: { type: number, exclusiveMinimum: 0 }
    Site:
      type: object
      additionalProperties: false
      required: [id, portfolio_id, name, timezone, grid_connection_limit_kw, status, created_at]
      properties:
        id: { type: string, format: uuid }
        portfolio_id: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string }
        grid_connection_limit_kw: { type: number }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        created_at: { type: string, format: date-time }
    CreateDeviceModelRequest:
      type: object
      additionalProperties: false
      required: [name, type, schema_version, point_schema]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        type: { type: string, enum: [METER, PV_INVERTER, BATTERY, EV_CHARGER] }
        schema_version: { type: integer, minimum: 1 }
        point_schema: { type: object, minProperties: 1, additionalProperties: true }
    DeviceModel:
      type: object
      additionalProperties: false
      required: [id, name, type, schema_version, point_schema, status, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string }
        schema_version: { type: integer }
        point_schema: { type: object, additionalProperties: true }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        created_at: { type: string, format: date-time }
    DeviceCapabilityInput:
      type: object
      additionalProperties: false
      required: [capability, unit, min_value, max_value, fallback_value]
      properties:
        capability: { type: string, minLength: 1, maxLength: 80 }
        unit: { type: string, minLength: 1, maxLength: 24 }
        min_value: { type: number }
        max_value: { type: number }
        fallback_value: { type: number }
    DeviceCapability:
      allOf:
        - $ref: "#/components/schemas/DeviceCapabilityInput"
        - type: object
          required: [config_version]
          properties:
            config_version: { type: integer, minimum: 1 }
    CreateDeviceRequest:
      type: object
      additionalProperties: false
      required: [site_id, model_id, external_code, name, capabilities]
      properties:
        site_id: { type: string, format: uuid }
        model_id: { type: string, format: uuid }
        external_code: { type: string, pattern: "^[a-zA-Z0-9][a-zA-Z0-9._-]{0,63}$" }
        name: { type: string, minLength: 1, maxLength: 120 }
        capabilities:
          type: array
          maxItems: 32
          items: { $ref: "#/components/schemas/DeviceCapabilityInput" }
    Device:
      type: object
      additionalProperties: false
      required: [id, site_id, model_id, external_code, name, status, config_version, credential_configured, capabilities, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        site_id: { type: string, format: uuid }
        model_id: { type: string, format: uuid }
        external_code: { type: string }
        name: { type: string }
        status: { type: string, enum: [PENDING, ACTIVE, MAINTENANCE, DISABLED, SAFETY_LOCKED] }
        config_version: { type: integer, minimum: 1 }
        credential_configured: { type: boolean }
        capabilities:
          type: array
          items: { $ref: "#/components/schemas/DeviceCapability" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UpdateDeviceStatusRequest:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { type: string, enum: [ACTIVE, MAINTENANCE, DISABLED, SAFETY_LOCKED] }
        reason: { type: string, minLength: 1, maxLength: 500 }
    RotateCredentialRequest:
      type: object
      additionalProperties: false
      required: [credential, reason]
      properties:
        credential: { type: string, minLength: 16, maxLength: 512, writeOnly: true }
        reason: { type: string, minLength: 1, maxLength: 500 }
    CredentialReference:
      type: object
      additionalProperties: false
      required: [id, device_id, status, rotated_at]
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, REVOKED] }
        rotated_at: { type: string, format: date-time }
    TariffIntervalInput:
      type: object
      additionalProperties: false
      required: [start, end, price_per_kwh]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        price_per_kwh: { type: number, minimum: 0 }
    CreateTariffPlanRequest:
      type: object
      additionalProperties: false
      required: [name, currency, timezone, version, valid_from, valid_to, intervals]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        timezone: { type: string, minLength: 1, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        valid_from: { type: string, format: date-time }
        valid_to: { type: string, format: date-time }
        intervals:
          type: array
          minItems: 1
          maxItems: 10000
          items: { $ref: "#/components/schemas/TariffIntervalInput" }
    TariffPlan:
      type: object
      additionalProperties: false
      required: [id, name, currency, timezone, version, valid_from, valid_to, interval_count, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        currency: { type: string }
        timezone: { type: string }
        version: { type: integer }
        valid_from: { type: string, format: date-time }
        valid_to: { type: string, format: date-time }
        interval_count: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
...
    CreateScheduleRequest:
      type: object
      additionalProperties: false
      required: [portfolio_id, schedule_date, load_forecast_version_id, pv_forecast_version_id, tariff_plan_id, battery_states]
      properties:
        portfolio_id: { type: string, format: uuid }
        schedule_date: { type: string, format: date }
        load_forecast_version_id: { type: string, format: uuid }
        pv_forecast_version_id: { type: string, format: uuid }
        tariff_plan_id: { type: string, format: uuid }
        battery_states:
          type: array
          maxItems: 256
          items:
            type: object
            additionalProperties: false
            required: [device_id, initial_soc_pct]
            properties:
              device_id: { type: string, format: uuid }
              initial_soc_pct: { type: number, minimum: 0, maximum: 100 }
    Schedule:
      type: object
      additionalProperties: false
      required: [id, portfolio_id, schedule_date, timezone, status, feasibility, reasons, input, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [VALIDATED, FAILED]
        feasibility: { type: string, enum: [FEASIBLE, INFEASIBLE, FAILED] }
        portfolio_id: { type: string, format: uuid }
        schedule_date: { type: string, format: date }
        timezone: { type: string }
        current_version: { type: [integer, "null"], minimum: 1 }
        failure_code: { type: [string, "null"] }
        reasons: { type: array, items: { type: string } }
        input: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ScheduleDetail:
      type: object
      additionalProperties: false
      required: [schedule, version]
      properties:
        schedule: { $ref: "#/components/schemas/Schedule" }
        version:
          type: [object, "null"]
          additionalProperties: true
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, trace_id]
      properties:
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        trace_id: { type: string, pattern: "^[a-f0-9]{32}$" }
        details:
          type: array
          items:
            type: object
            additionalProperties: true
            required: [reason]
            properties:
              field: { type: string }
              reason: { type: string }
security:
  - bearerAuth: []